changed LineGraph to SingleLineGraph
